Extract service registrations from TendrilServer into ServiceRegistration.cs
Moves all DI container setup into a dedicated extension method, reducing
TendrilServer.Create() from 215 LoC/cc=14 to 79 LoC. Code health improves
from 8.85 to 9.55.
